rename graphviz.py in order to avoid conflict with the python graphviz lib. test=develop
